+++ lib/signalwire/sdk.rb
@@ -10,6 +10,7 @@ require 'signalwire/sdk/voice_response'
 require 'signalwire/sdk/fax_response'
 require 'signalwire/sdk/messaging_response'
 require 'signalwire/rest/client'
+require 'signalwire/webhook/validate_request'
 
 module Signalwire
   module Sdk